Move tags_from_query_string to core.utils
